Add files via upload
The project contains two problems:

PROBLEM 1 : 
Your friend is a chief architect, who is working on building a skyscraper, in Mumbai. The
construction is in such a way that the floors will be constructed in other factories and they will be
assembled. All the sizes will be distinct.
The skyscraper needs to be constructed in N days with the following conditions :
a) Every day a floor is constructed in a separate factory of distinct size.
b) The floor with the larger size must be placed at the bottom of the building.
c) The floor with the smaller size must be placed at the top of the building.

PROBLEM 2:  
You are working in an MNC, which manages the Transactions, where only BST is used as a
Data Structure. The company stores all the data of transactions in BST such that the tree is
always a complete BST.
A new business requirement has arrived where the BST should not contain any left node.
You are required to modify the existing BST and display the node values present in ascending
order.